requirements.txt
numpy==1.18.1
opencv-contrib-python-nonfree==4.1.1.1
Pillow==6.2.2
scikit-learn==0.22.1
scipy==1.4.1
simplejson==3.13.2
six==1.14.0
sklearn==0.0
torch==1.3.1
torchvision==0.4.2
